+++ examples/10_threading.rs
@@ -10,7 +10,7 @@ use afire::{Header, Method, Response, Server};
 
 fn main() {
     // Create a new Server instance on localhost port 8080
-    let mut server: Server = Server::new("localhost", 1313);
+    let mut server: Server = Server::new("localhost", 8080);
 
     // Define a handler for GET "/"
     server.route(Method::GET, "/", |_req| {